move ModalChecklist to components
